+++ handlers/__init__.py
@@ -1,13 +1,13 @@
 from .group_handler import GroupHandler
 from .join_handler import JoinHandler
 from .message_handler import MessageHandler
-from .sentinel_handler import SentinelHandler
+from .sentinel_handler import InspectionHandler
 from .time_parser import TimeParser
 
 __all__ = [
     "GroupHandler",
+    "InspectionHandler",
     "JoinHandler",
     "MessageHandler",
-    "SentinelHandler",
     "TimeParser",
 ]